[ISSUE #5081] Enhancement update for connectors & admin-server
