+++ eventmesh-common/src/main/java/org/apache/eventmesh/common/config/ConfigService.java
@@ -131,7 +131,7 @@ public class ConfigService {
         } else {
             filePath = path.startsWith(FILE_PATH_PREFIX) ? path.substring(FILE_PATH_PREFIX.length()) : this.configPath + path;
         }
-
+        filePath = normalizeFilePath(filePath);
         if (filePath.contains(".jar")) {
             try (final InputStream inputStream = getClass().getResourceAsStream(Objects.requireNonNull(resourceUrl))) {
                 if (inputStream == null) {
@@ -150,6 +150,15 @@ public class ConfigService {
         configInfo.setResourceUrl(resourceUrl);
         object = FileLoad.getFileLoad(suffix).getConfig(configInfo);
         return (T) object;
+    }
+
+    private String normalizeFilePath(String filePath) {
+        if (System.getProperty("os.name").toLowerCase().contains("win")) {
+            if (filePath.startsWith("/")) {
+                filePath = filePath.substring(1);
+            }
+        }
+        return filePath;
     }
 
     private void populateConfig(Object object, Class<?> clazz, Config config)
