+++ eventmesh-common/src/main/java/org/apache/eventmesh/common/config/connector/rdb/canal/CanalSinkConfig.java
@@ -18,7 +18,8 @@
 package org.apache.eventmesh.common.config.connector.rdb.canal;
 
 import org.apache.eventmesh.common.config.connector.SinkConfig;
-import org.apache.eventmesh.common.remote.job.SyncMode;
+
+import java.util.Map;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -27,25 +28,7 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = true)
 public class CanalSinkConfig extends SinkConfig {
 
-    // batchSize
-    private Integer batchSize = 50;
-
-    // enable batch
-    private Boolean useBatch = true;
-
-    // sink thread size for single channel
-    private Integer poolSize = 5;
-
-    // sync mode: field/row
-    private SyncMode syncMode;
-
-    private boolean isGTIDMode = true;
-
-    private boolean isMariaDB = true;
-
-    // skip sink process exception
-    private Boolean skipException = false;
-
-    public SinkConnectorConfig sinkConnectorConfig;
+    // used to convert canal full/increment/check connector config
+    private Map<String, Object> sinkConfig;
 
 }
