+++ eventmesh-connectors/eventmesh-connector-canal/src/main/java/org/apache/eventmesh/connector/canal/source/connector/CanalSourceFullConnector.java
@@ -20,6 +20,7 @@ package org.apache.eventmesh.connector.canal.source.connector;
 import org.apache.eventmesh.common.AbstractComponent;
 import org.apache.eventmesh.common.EventMeshThreadFactory;
 import org.apache.eventmesh.common.config.connector.Config;
+import org.apache.eventmesh.common.config.connector.rdb.canal.CanalSourceConfig;
 import org.apache.eventmesh.common.config.connector.rdb.canal.CanalSourceFullConfig;
 import org.apache.eventmesh.common.config.connector.rdb.canal.JobRdbFullPosition;
 import org.apache.eventmesh.common.config.connector.rdb.canal.RdbDBDefinition;
@@ -32,11 +33,11 @@ import org.apache.eventmesh.connector.canal.source.position.CanalFullPositionMgr
 import org.apache.eventmesh.connector.canal.source.position.TableFullPosition;
 import org.apache.eventmesh.connector.canal.source.table.RdbSimpleTable;
 import org.apache.eventmesh.connector.canal.source.table.RdbTableMgr;
-import org.apache.eventmesh.openconnect.api.ConnectorCreateService;
 import org.apache.eventmesh.openconnect.api.connector.ConnectorContext;
 import org.apache.eventmesh.openconnect.api.connector.SourceConnectorContext;
 import org.apache.eventmesh.openconnect.api.source.Source;
 import org.apache.eventmesh.openconnect.offsetmgmt.api.data.ConnectRecord;
+import org.apache.eventmesh.openconnect.util.ConfigUtil;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -49,7 +50,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class CanalSourceFullConnector extends AbstractComponent implements Source, ConnectorCreateService<Source> {
+public class CanalSourceFullConnector extends AbstractComponent implements Source {
+
     private CanalSourceFullConfig config;
     private CanalFullPositionMgr positionMgr;
     private RdbTableMgr tableMgr;
@@ -62,14 +64,14 @@ public class CanalSourceFullConnector extends AbstractComponent implements Sourc
         this.tableMgr.start();
         this.positionMgr.start();
         if (positionMgr.isFinished()) {
-            log.info("connector [{}] has finished the job", config.getConnectorConfig().getConnectorName());
+            log.info("connector [{}] has finished the job", config.getSourceConnectorConfig().getConnectorName());
             return;
         }
         executor = new ThreadPoolExecutor(config.getParallel(), config.getParallel(), 0L, TimeUnit.MILLISECONDS,
             new LinkedBlockingQueue<>(), new EventMeshThreadFactory("canal-source-full"));
         List<CanalFullProducer> producers = new LinkedList<>();
-        if (config.getConnectorConfig().getDatabases() != null) {
-            for (RdbDBDefinition db : config.getConnectorConfig().getDatabases()) {
+        if (config.getSourceConnectorConfig().getDatabases() != null) {
+            for (RdbDBDefinition db : config.getSourceConnectorConfig().getDatabases()) {
                 for (RdbTableDefinition table : db.getTables()) {
                     try {
                         log.info("it will create producer of db [{}] table [{}]", db.getSchemaName(), table.getTableName());
@@ -120,11 +122,6 @@ public class CanalSourceFullConnector extends AbstractComponent implements Sourc
     }
 
     @Override
-    public Source create() {
-        return new CanalSourceFullConnector();
-    }
-
-    @Override
     public Class<? extends Config> configClass() {
         return CanalSourceFullConfig.class;
     }
@@ -136,16 +133,17 @@ public class CanalSourceFullConnector extends AbstractComponent implements Sourc
     }
 
     private void init() {
-        DatabaseConnection.sourceConfig = this.config.getConnectorConfig();
+        DatabaseConnection.sourceConfig = this.config.getSourceConnectorConfig();
         DatabaseConnection.initSourceConnection();
-        this.tableMgr = new RdbTableMgr(config.getConnectorConfig(), DatabaseConnection.sourceDataSource);
+        this.tableMgr = new RdbTableMgr(config.getSourceConnectorConfig(), DatabaseConnection.sourceDataSource);
         this.positionMgr = new CanalFullPositionMgr(config, tableMgr);
     }
 
     @Override
     public void init(ConnectorContext connectorContext) throws Exception {
         SourceConnectorContext sourceConnectorContext = (SourceConnectorContext) connectorContext;
-        this.config = (CanalSourceFullConfig) sourceConnectorContext.getSourceConfig();
+        CanalSourceConfig canalSourceConfig = (CanalSourceConfig) sourceConnectorContext.getSourceConfig();
+        this.config = ConfigUtil.parse(canalSourceConfig.getSourceConfig(), CanalSourceFullConfig.class);
         init();
     }
 
@@ -156,7 +154,7 @@ public class CanalSourceFullConnector extends AbstractComponent implements Sourc
 
     @Override
     public String name() {
-        return this.config.getConnectorConfig().getConnectorName();
+        return this.config.getSourceConnectorConfig().getConnectorName();
     }
 
     @Override
