+++ eventmesh-connectors/eventmesh-connector-http/src/main/java/org/apache/eventmesh/connector/http/sink/handler/HttpSinkHandler.java
@@ -33,14 +33,14 @@ import io.vertx.ext.web.client.HttpResponse;
  *
  * <p>Any class that needs to process ConnectRecords via HTTP or HTTPS should implement this interface.
  * Implementing classes must provide implementations for the {@link #start()}, {@link #handle(ConnectRecord)},
- * {@link #deliver(URI, HttpConnectRecord, Map)}, and {@link #stop()} methods.</p>
+ * {@link #deliver(URI, HttpConnectRecord, Map, ConnectRecord)}, and {@link #stop()} methods.</p>
  *
  * <p>Implementing classes should ensure thread safety and handle HTTP/HTTPS communication efficiently.
  * The {@link #start()} method initializes any necessary resources for HTTP/HTTPS communication. The {@link #handle(ConnectRecord)} method processes a
- * ConnectRecord by sending it over HTTP or HTTPS. The {@link #deliver(URI, HttpConnectRecord, Map)} method processes HttpConnectRecord on specified
- * URL while returning its own processing logic {@link #stop()} method releases any resources used for HTTP/HTTPS communication.</p>
+ * ConnectRecord by sending it over HTTP or HTTPS. The {@link #deliver(URI, HttpConnectRecord, Map, ConnectRecord)} method processes HttpConnectRecord
+ * on specified URL while returning its own processing logic {@link #stop()} method releases any resources used for HTTP/HTTPS communication.</p>
  *
- * <p>It's recommended to handle exceptions gracefully within the {@link #deliver(URI, HttpConnectRecord, Map)} method
+ * <p>It's recommended to handle exceptions gracefully within the {@link #deliver(URI, HttpConnectRecord, Map, ConnectRecord)} method
  * to prevent message loss or processing interruptions.</p>
  */
 public interface HttpSinkHandler {
@@ -66,7 +66,7 @@ public interface HttpSinkHandler {
      * @param attributes        additional attributes to be used in processing
      * @return processing chain
      */
-    Future<HttpResponse<Buffer>> deliver(URI url, HttpConnectRecord httpConnectRecord, Map<String, Object> attributes);
+    Future<HttpResponse<Buffer>> deliver(URI url, HttpConnectRecord httpConnectRecord, Map<String, Object> attributes, ConnectRecord connectRecord);
 
     /**
      * Cleans up and releases resources used by the HTTP/HTTPS handler. This method should be called when the handler is no longer needed.
