+++ eventmesh-openconnect/eventmesh-openconnect-java/src/main/java/org/apache/eventmesh/openconnect/api/connector/SourceConnectorContext.java
@@ -18,6 +18,7 @@
 package org.apache.eventmesh.openconnect.api.connector;
 
 import org.apache.eventmesh.common.config.connector.SourceConfig;
+import org.apache.eventmesh.common.remote.job.JobType;
 import org.apache.eventmesh.common.remote.offset.RecordPosition;
 import org.apache.eventmesh.openconnect.offsetmgmt.api.storage.OffsetStorageReader;
 
@@ -37,6 +38,8 @@ public class SourceConnectorContext implements ConnectorContext {
     public SourceConfig sourceConfig;
 
     public Map<String, Object> runtimeConfig;
+
+    public JobType jobType;
 
     // initial record position
     public List<RecordPosition> recordPositionList;
